+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -22,7 +22,7 @@ use Symfony\Component\Config\Definition\Processor;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HtmlSanitizer\HtmlSanitizer;
 use Symfony\Component\HttpClient\HttpClient;
-use Symfony\Component\JsonEncoder\JsonEncoder;
+use Symfony\Component\JsonStreamer\JsonStreamWriter;
 use Symfony\Component\Lock\Store\SemaphoreStore;
 use Symfony\Component\Mailer\Mailer;
 use Symfony\Component\Messenger\MessageBusInterface;
@@ -1009,8 +1009,8 @@ class ConfigurationTest extends TestCase
             'remote-event' => [
                 'enabled' => !class_exists(FullStack::class) && class_exists(RemoteEvent::class),
             ],
-            'json_encoder' => [
-                'enabled' => !class_exists(FullStack::class) && class_exists(JsonEncoder::class),
+            'json_streamer' => [
+                'enabled' => !class_exists(FullStack::class) && class_exists(JsonStreamWriter::class),
             ],
         ];
     }