+++ src/Symfony/Component/JsonEncoder/Tests/Fixtures/Model/DummyWithDateTimes.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Symfony\Component\JsonEncoder\Tests\Fixtures\Model;
-
-class DummyWithDateTimes
-{
-    public \DateTimeInterface $interface;
-    public \DateTimeImmutable $immutable;
-}